V5 (#60)
# V5 - Daterange completeness, dynamic resizability!

Datepicker now handles daterange pairs completely. No more having to provide your own `onChange` function to each instance in the pair. Everything is handled under the hood. The other major change is that the calendar can be dynamically sized via the font-size property! The calendar is within a new container div which has the font-size property set on it. Change that with JavaScript and vóila! The calendar has a new size.

Here's a list of everything changed in this version:

- Today's date is now underlined and always bold (even when disabled).
- Moved from using Less to using SCSS!
- All callbacks get a default noop function value.
- Daterange pairs are now completely handled internally. They work out of the box. No more having to provide custom logic in your callback functions.
- Datepicker is now completely dynamically resizable! Added the calendar to a container element which can have a font-size which controls the calendars size. All calendar elements are relatively sized to support this.
- Added a getRange method for daterange pairs.
- Moved index.ejs to the sandbox folder.