Added Cypress e2e tests; re-wrote library in es5 (#77)
## Cypress.io e2e Testing
The library is now tested completely with Cypress.io! This will help abstract away any implementation details and test the library as a user would _use_ it. This was a long time in the making. Just had to find the time to complete it. I removed the previous half-implemented Jest tests. No love lost for Jest :)

## ES5 (or 3?)
* I re-wrote the entire library in ES5 syntax. This should solve all the issues on older browsers such as IE11. Maybe even down to IE9? Only time will tell...
* To help with ensuring the syntax was ES5, I added eslint to the project and set the ecma version to 3.
* As a result of the ES5 re-write, the webpack config was tweaked to no longer use babel & friends. In fact, all it's doing for JS is minification.